+++ wcomponents-core/src/main/java/com/github/bordertech/wcomponents/render/webxml/WDropdownRenderer.java
@@ -90,6 +90,9 @@ final class WDropdownRenderer extends AbstractWebXmlRenderer {
 			}
 		}
 
+		if (!readOnly) {
+			DiagnosticRenderUtil.renderDiagnostics(dropdown, renderContext);
+		}
 		// End tag
 		xml.appendEndTag("ui:dropdown");
 	}