+++ wcomponents-core/src/test/java/com/github/bordertech/wcomponents/validation/DiagnosticImpl_Test.java
@@ -58,6 +58,10 @@ public class DiagnosticImpl_Test {
 		Assert.assertEquals("Incorrect description text", "The field 'a' is required", diag.
 				getDescription());
 
+		// Test with formatting with toolTip
+		input.setAccessibleText(null);
+		input.setToolTip("a tooltip");
+		Assert.assertEquals("Incorrect description text", "The field 'a tooltip' is required", diag.getDescription());
 		// Test with label set
 		WLabel label = new WLabel("bc", input);
 		Assert.assertEquals("Incorrect description text", "The field 'bc' is required", diag.