+++ wcomponents-core/src/test/java/com/github/bordertech/wcomponents/validation/WFieldErrorIndicator_Test.java
@@ -5,7 +5,7 @@ import com.github.bordertech.wcomponents.WPanel;
 import com.github.bordertech.wcomponents.WTextField;
 import java.util.ArrayList;
 import java.util.List;
-import junit.framework.Assert;
+import org.junit.Assert;
 import org.junit.Test;
 
 /**
@@ -20,14 +20,9 @@ public class WFieldErrorIndicator_Test extends AbstractWComponentTestCase {
 	public void testConstructor() {
 		WTextField component = new WTextField();
 		WFieldErrorIndicator indicator = new WFieldErrorIndicator(component);
-
-		Assert.assertEquals("Incorrect indicator type",
-				AbstractWFieldIndicator.FieldIndicatorType.ERROR, indicator
-				.getFieldIndicatorType());
+		Assert.assertEquals("Incorrect indicator type", 	AbstractWFieldIndicator.FieldIndicatorType.ERROR, indicator.getFieldIndicatorType());
 		Assert.assertEquals("Incorrect releated field", component, indicator.getRelatedField());
-
-		Assert.assertEquals("Incorrect releated field id", component.getId(), indicator.
-				getRelatedFieldId());
+		Assert.assertEquals("Incorrect releated field id", component.getId(), indicator.getRelatedFieldId());
 	}
 
 	@Test
@@ -55,8 +50,7 @@ public class WFieldErrorIndicator_Test extends AbstractWComponentTestCase {
 		root.validate(diags);
 		root.showErrorIndicators(diags);
 
-		Assert.assertFalse("Should not be in default if there are errors", indicator.
-				isDefaultState());
+		Assert.assertFalse("Should not be in default if there are errors", indicator.isDefaultState());
 
 		root.reset();
 		Assert.assertTrue("Should be in default after reset", indicator.isDefaultState());