+++ wcomponents-theme/src/main/js/wc/ui/checkBox.js
@@ -1,22 +1,9 @@
-/**
- * Provides Ajax and state writing functionality for check boxes.
- * @module
- * @requires module:wc/dom/initialise
- * @requires module:wc/dom/Widget
- * @requires module:wc/dom/shed
- * @requires module:wc/dom/formUpdateManager
- * @requires module:wc/ui/ajax/processResponse
- * @requires module:wc/dom/cbrShedPublisher
- *
- * @todo Get rid of the state writing: it is nuts!
- */
 define(["wc/dom/initialise",
 	"wc/dom/Widget",
 	"wc/dom/shed",
 	"wc/dom/formUpdateManager",
 	"wc/ui/ajax/processResponse",
 	"wc/dom/cbrShedPublisher"],
-	/** @param initialise wc/dom/initialise @param Widget wc/dom/Widget @param shed wc/dom/shed @param formUpdateManager wc/dom/formUpdateManager @param processResponse wc/ui/ajax/processResponse @ignore */
 	function(initialise, Widget, shed, formUpdateManager, processResponse) {
 		"use strict";
 
@@ -27,17 +14,29 @@ define(["wc/dom/initialise",
 		 */
 		function CheckBox() {
 			var CHECKBOX = new Widget("input", "", { "type": "checkbox" }),
-				WRAPPER,
-				CB_ALONE;
+				WRAPPER = new Widget("", "wc-checkbox"),
+				WCHECKBOX = CHECKBOX.clone();
 
+			WCHECKBOX.descendFrom(WRAPPER, true);
 			/**
 			 * Provides the {@link module:wc/dom/Widget} description of a CHECKBOX.
 			 * @function module:wc/ui/checkbox.getWidget
 			 * @public
-			 * @returns {Widget}
+			 * @param {boolean} [onlyWcb] if `true` return the Widget to explicitly match WCheckBox
+			 * @returns {checkBoxL#7.Widget} the description of a CHECKBOX; or WCHECKBOX is onlyWcb is `truthy`.
 			 */
-			this.getWidget = function() {
-				return CHECKBOX;
+			this.getWidget = function(onlyWcb) {
+				return onlyWcb ? WCHECKBOX : CHECKBOX;
+			};
+
+			/**
+			 * Provides the {@link module:wc/dom/Widget} description of a CHECKBOXWCheckBox wrapper element
+			 * @function module:wc/ui/checkbox.getWrapper
+			 * @public
+			 * @returns {checkBoxL#7.Widget}
+			 */
+			this.getWrapper = function() {
+				return WRAPPER;
 			};
 
 			/**
@@ -54,10 +53,10 @@ define(["wc/dom/initialise",
 				var cb;
 				if (!WRAPPER) {
 					WRAPPER = new Widget("", "wc-checkbox");
-					CB_ALONE = CHECKBOX.clone();
-					CB_ALONE.descendFrom(WRAPPER, true);
+					WCHECKBOX = CHECKBOX.clone();
+					WCHECKBOX.descendFrom(WRAPPER, true);
 				}
-				cb = CB_ALONE.findDescendants(form);
+				cb = WCHECKBOX.findDescendants(form);
 
 				cb = Array.prototype.filter.call(cb, function (next) {
 					return !(shed.isSelected(next) || shed.isDisabled(next));
@@ -127,7 +126,22 @@ define(["wc/dom/initialise",
 			};
 		}
 
-		var /** @alias module:wc/ui/checkbox */ instance = new CheckBox();
+		/**
+		 * Provides Ajax and state writing functionality for check boxes.
+		 *
+		 * {@link module:wc/dom/cbrShedPublisher} is listed as a dependency as its functionality is required by native checkbox elements but it is not
+		 * used by the code in this module.
+		 *
+		 * @module
+		 * @requires wc/dom/initialise
+		 * @requires wc/dom/Widget
+		 * @requires wc/dom/shed
+		 * @requires wc/dom/formUpdateManager
+		 * @requires wc/ui/ajax/processResponse
+		 *
+		 * @todo Get rid of the state writing: it is nuts!
+		 */
+		var instance = new CheckBox();
 		initialise.register(instance);
 		return instance;
 	});