+++ wcomponents-theme/src/main/js/wc/ui/radioButtonSelect.js
@@ -17,7 +17,7 @@ define(["wc/dom/group",
 		 */
 		function RadioButtonGroup() {
 			var RADIO = new Widget("input", null, {"type": "radio"}),
-				RADIO_BUTTON_SELECT = new Widget("", "wc-radioButtonSelect", {"role": "radiogroup"});
+				RADIO_BUTTON_SELECT = new Widget("fieldset", "wc-radiobuttonselect");
 
 			/**
 			 * Listen for mandatory/optional and set the group's radio buttons.